requirements.txt
fastapi
uvicorn
playwright
jinja2
python-dotenv
aiohttp
pydantic
